requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Video Processing
moviepy==1.0.3
imageio==2.31.5
imageio-ffmpeg==0.4.9
Pillow==10.1.0

# AI Models
google-generativeai==0.3.2
transformers==4.35.2
torch==2.1.1
diffusers==0.24.0
accelerate==0.25.0

# Replicate (Alternative)
replicate==0.22.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
python-jose[cryptography]==3.3.0

# Frontend
streamlit==1.28.2

# Job Queue (Optional - for production)
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

